refactor(api): clean up split operation
