+++ api/apps/api/src/modules/scenarios-features/create-features.handler.ts
@@ -5,17 +5,12 @@ import {
 } from '@nestjs/cqrs';
 import { EntityManager } from 'typeorm';
 import { InjectEntityManager } from '@nestjs/typeorm';
-import {
-  FeatureConfigSplit,
-  SpecificationOperation,
-} from '@marxan-api/modules/specification';
-import { Scenario } from '@marxan-api/modules/scenarios/scenario.api.entity';
-import { ProjectsCrudService } from '@marxan-api/modules/projects/projects-crud.service';
+import { SpecificationOperation } from '@marxan-api/modules/specification';
 import { DbConnections } from '@marxan-api/ormconfig.connections';
-import { assertDefined, isDefined } from '@marxan/utils';
 import { CreateFeaturesCommand } from './create-features.command';
 import { FeaturesCreated } from './features-created.event';
 import { CopyOperation } from './copy';
+import { SplitOperation } from './split';
 
 @CommandHandler(CreateFeaturesCommand)
 export class CreateFeaturesHandler
@@ -23,11 +18,9 @@ export class CreateFeaturesHandler
   constructor(
     private readonly eventBus: EventBus,
     private readonly copyOperation: CopyOperation,
-    @InjectEntityManager()
-    private readonly apiEntityManager: EntityManager,
+    private readonly splitOperation: SplitOperation,
     @InjectEntityManager(DbConnections.geoprocessingDB)
     private readonly geoEntityManager: EntityManager,
-    private readonly projects: ProjectsCrudService,
   ) {}
 
   async execute(command: CreateFeaturesCommand): Promise<void> {
@@ -36,7 +29,10 @@ export class CreateFeaturesHandler
         break;
       }
       case SpecificationOperation.Copy: {
-        const ids = await this.copyOperation.copy(command);
+        const ids = await this.copyOperation.copy({
+          ...command,
+          input: command.input,
+        });
         this.eventBus.publish(
           new FeaturesCreated(command.scenarioId, {
             ...command.input,
@@ -46,126 +42,10 @@ export class CreateFeaturesHandler
         break;
       }
       case SpecificationOperation.Split: {
-        const scenario = await this.apiEntityManager
-          .getRepository(Scenario)
-          .findOne(command.scenarioId, {
-            relations: ['project'],
-          });
-        assertDefined(scenario);
-        const { project } = scenario;
-        assertDefined(project);
-        const input = command.input as FeatureConfigSplit;
-
-        const protectedAreaFilterByIds =
-          scenario.protectedAreaFilterByIds ?? [];
-
-        const planningAreaLocation = await this.projects.locatePlanningAreaEntity(
-          {
-            adminAreaLevel1Id: project.adminAreaLevel1Id,
-            adminAreaLevel2Id: project.adminAreaLevel2Id,
-            countryId: project.countryId,
-            planningAreaGeometryId: project.planningAreaGeometryId,
-          },
-        );
-
-        const hasSubSetFilter = (input.selectSubSets ?? []).length > 0;
-
-        const parameters: (string | number)[] = [];
-        const fields = {
-          filterByValues: (input.selectSubSets ?? []).map(
-            (s) => `$${parameters.push(s.value)}`,
-          ),
-          splitByProperty: `$${parameters.push(input.splitByProperty)}`,
-          scenarioId: `$${parameters.push(command.scenarioId)}`,
-          planningAreaId: isDefined(planningAreaLocation)
-            ? `$${parameters.push(planningAreaLocation.id)}`
-            : `NULL`,
-          protectedAreaIds:
-            protectedAreaFilterByIds.length > 0
-              ? protectedAreaFilterByIds
-                  .map((id) => `$${parameters.push(id)}`)
-                  .join(', ')
-              : undefined,
-          protectedArea:
-            protectedAreaFilterByIds.length > 0 ? 'protected.area' : 'NULL',
-          featureId: `$${parameters.push(command.input.baseFeatureId)}`,
-          bbox: [
-            `$${parameters.push(project.bbox[0])}`,
-            `$${parameters.push(project.bbox[1])}`,
-            `$${parameters.push(project.bbox[2])}`,
-            `$${parameters.push(project.bbox[3])}`,
-          ],
-          totalArea: isDefined(planningAreaLocation)
-            ? `st_area(st_intersection(pa.the_geom, fd.the_geom))`
-            : `NULL`,
-        };
-        const protectedAreaJoin = fields.protectedAreaIds
-          ? `cross join (
-                   select st_area(st_union(wdpa.the_geom)) as area
-                   from wdpa where wdpa.id in (${fields.protectedAreaIds})
-                 ) as protected`
-          : '';
-        const planningAreaJoin = isDefined(planningAreaLocation)
-          ? `left join ${planningAreaLocation.tableName} as pa on pa.id = ${fields.planningAreaId}`
-          : ``;
-        // const propertiesJoin = `left join (json_to_recordset???) as props on (props.id = fd.id)`;
-
-        const ids: { id: string }[] = await this.geoEntityManager.query(
-          `
-            insert into scenario_features_data as sfd (feature_class_id,
-                                                     scenario_id,
-                                                     fpf,
-                                                     target,
-                                                     prop,
-                                                     total_area,
-                                                     current_pa)
-            WITH split as (
-              WITH subsets as (
-              select value as sub_value, target, fpf, prop
-              from json_to_recordset('${JSON.stringify(
-                input.selectSubSets ?? [],
-              )}') as x(value varchar, target float8, fpf float8,
-              prop float8)
-              )
-              SELECT distinct fpkv.feature_data_id,
-              fpkv.key,
-              fpkv.value,
-              subsets.fpf,
-              subsets.prop,
-              subsets.target
-              from feature_properties_kv fpkv
-              left join subsets
-              on (('[' || to_json(fpkv.value::TEXT) || ']')::json ->> 0 = subsets.sub_value:: varchar)
-              where fpkv.feature_id = ${fields.featureId}
-              and fpkv.key = ${fields.splitByProperty} ${
-            hasSubSetFilter
-              ? `and fpkv.value IN(${fields.filterByValues.join(',')})`
-              : ``
-          }
-              )
-            select fd.id,
-                   ${fields.scenarioId},
-                   split.fpf,
-                   split.target,
-                   split.prop,
-                   ${fields.totalArea},
-                   ${fields.protectedArea}
-            from split
-                   join features_data as fd
-                        on (split.feature_data_id = fd.id)
-              ${planningAreaJoin} ${protectedAreaJoin}
-
-            where st_intersects(st_makeenvelope( ${fields.bbox[0]}
-                , ${fields.bbox[1]}
-                , ${fields.bbox[2]}
-                , ${fields.bbox[3]}
-                , 4326
-              )
-                , fd.the_geom)
-              returning sfd.id as id;
-          `,
-          parameters,
-        );
+        const ids = await this.splitOperation.split({
+          ...command,
+          input: command.input,
+        });
         this.eventBus.publish(
           new FeaturesCreated(command.scenarioId, {
             ...command.input,
