+++ api/apps/api/src/modules/scenarios-features/split/index.ts
@@ -1,0 +1,3 @@
+export { SplitDataProvider } from './split-data.provider';
+export { SplitOperation } from './split-operation.service';
+export { SplitQuery } from './split-query.service';
