+++ api/apps/api/src/modules/scenarios-features/split/split-operation.service.ts
@@ -1,38 +1,42 @@
-import { Injectable } from '@nestjs/common';
 import { EntityManager } from 'typeorm';
 import { InjectEntityManager } from '@nestjs/typeorm';
+import { Injectable } from '@nestjs/common';
 import { API_EVENT_KINDS } from '@marxan/api-events';
 import { DbConnections } from '@marxan-api/ormconfig.connections';
+import { FeatureConfigSplit } from '@marxan-api/modules/specification';
 import { ApiEventsService } from '@marxan-api/modules/api-events/api-events.service';
-import { FeatureConfigCopy } from '@marxan-api/modules/specification';
 import { CreateFeaturesCommand } from '../create-features.command';
-import { CopyQuery } from './copy-query.service';
-import { CopyDataProvider } from './copy-data-provider.service';
+import { SplitDataProvider } from './split-data.provider';
+import { SplitQuery } from './split-query.service';
 
 @Injectable()
-export class CopyOperation {
+export class SplitOperation {
   constructor(
-    private readonly copyQuery: CopyQuery,
-    private readonly copyDataProvider: CopyDataProvider,
+    private readonly splitDataProvider: SplitDataProvider,
+    private readonly splitQuery: SplitQuery,
     @InjectEntityManager(DbConnections.geoprocessingDB)
     private readonly geoEntityManager: EntityManager,
     private readonly events: ApiEventsService,
   ) {}
 
-  async copy(
-    command: CreateFeaturesCommand & { input: FeatureConfigCopy },
-  ): Promise<{ id: string }[]> {
+  async split(command: CreateFeaturesCommand & { input: FeatureConfigSplit }) {
     await this.events.create({
       topic: command.scenarioId,
-      kind: API_EVENT_KINDS.scenario__geofeatureCopy__submitted__v1__alpha1,
+      kind: API_EVENT_KINDS.scenario__geofeatureSplit__submitted__v1__alpha1,
     });
     try {
       const {
         project,
+        input,
         protectedAreaFilterByIds,
         planningAreaLocation,
-      } = await this.copyDataProvider.prepareData(command);
-      const { parameters, query } = this.copyQuery.prepareStatement(
+      } = await this.splitDataProvider.prepareData({
+        ...command,
+        input: command.input,
+      });
+
+      const { parameters, query } = this.splitQuery.prepareQuery(
+        input,
         command,
         planningAreaLocation,
         protectedAreaFilterByIds,
@@ -44,13 +48,13 @@ export class CopyOperation {
       );
       await this.events.create({
         topic: command.scenarioId,
-        kind: API_EVENT_KINDS.scenario__geofeatureCopy__finished__v1__alpha1,
+        kind: API_EVENT_KINDS.scenario__geofeatureSplit__finished__v1__alpha1,
       });
       return ids;
     } catch (error) {
       await this.events.create({
         topic: command.scenarioId,
-        kind: API_EVENT_KINDS.scenario__geofeatureCopy__failed__v1__alpha1,
+        kind: API_EVENT_KINDS.scenario__geofeatureSplit__failed__v1__alpha1,
       });
       throw error;
     }
