+++ api/apps/api/src/modules/scenarios/specification/geo-feature-dto.mapper.spec.ts
@@ -28,6 +28,7 @@ test(`split configuration`, () => {
       SpecForGeoFeatureWithGeoprocessing
     >(SpecForGeoFeatureWithGeoprocessing, {
       featureId,
+      kind: 'withGeoprocessing',
       geoprocessingOperations: [
         plainToClass<
           GeoprocessingOpStratificationV1,
