+++ api/apps/api/src/modules/specification/domain/__tests__/determine-features/fixtures.ts
@@ -30,12 +30,10 @@ export const getFixtures = () => {
     {
       featureId: v4(),
       calculated: true,
-      featureId: v4(),
     },
     {
       featureId: v4(),
       calculated: true,
-      featureId: v4(),
     },
   ];
 
@@ -43,12 +41,10 @@ export const getFixtures = () => {
     {
       featureId: v4(),
       calculated: true,
-      featureId: v4(),
     },
     {
       featureId: nonCalculatedFeature,
       calculated: false,
-      featureId: v4(),
     },
   ];
 
@@ -56,12 +52,10 @@ export const getFixtures = () => {
     {
       featureId: v4(),
       calculated: true,
-      featureId: v4(),
     },
     {
       featureId: v4(),
       calculated: true,
-      featureId: v4(),
     },
   ];
